refactor: move LoginScreenEvent to separate file
